third_party/WebKit/LayoutTests/accessibility/input-type-text-caret-position-expected.txt

This tests that caret position is reported correctly for text input elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS inputAccessible.selectionStart is 0
PASS inputAccessible.selectionEnd is 0
PASS inputAccessible.selectionStart is 1
PASS inputAccessible.selectionEnd is 1
PASS inputAccessible.selectionStart is 2
PASS inputAccessible.selectionEnd is 2
PASS inputAccessible.selectionStart is 3
PASS inputAccessible.selectionEnd is 3
PASS inputAccessible.selectionStart is 4
PASS inputAccessible.selectionEnd is 4
PASS inputAccessible.selectionStart is 5
PASS inputAccessible.selectionEnd is 5
PASS inputAccessible.selectionStart is 6
PASS inputAccessible.selectionEnd is 6
PASS successfullyParsed is true

TEST COMPLETE
